requirements.txt
Pillow
uvicorn
fastapi
requests
